agents/file_agent/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1

# PDF/Image + OCR
PyMuPDF==1.23.7
pytesseract>=0.3.13
Pillow>=11.3.0

# Redis and PubSub
redis==5.0.4

# NLP & ML
sentence-transformers==2.6.1
transformers>=4.32.0,<5.0.0
torch==2.7.1

huggingface-hub>=0.15.1

# Pinecone
pinecone-client==3.2.2

# Other utils
python-dotenv==1.0.1
orjson>=3.2.1
httpx>=0.23.0
python-multipart>=0.0.7
jinja2>=3.1.3
typing-extensions>=4.8.0
anyio<5,>=3.5.0
watchfiles>=0.13
typer>=0.15.1
rich>=13.7.1
